refactor: update generatePancakeSwapTokenRecords to use tsx and handlebars
