+++ package.json
@@ -22,7 +22,7 @@
     "generate-subgraph-types:testnet": "yarn graphql-codegen --config ./src/config/codegen.ts --project testnet",
     "generate-subgraph-types": "npm-run-all --parallel generate-subgraph-types:mainnet generate-subgraph-types:testnet",
     "generate-contracts": "rimraf ./src/packages/contracts/generated/getters && rimraf ./src/packages/contracts/generated/infos && ./src/packages/contracts/scripts/generateContractRecords/index.ts && yarn prettier --write \"src/packages/contracts/generated/getters/**/*.{ts,json}\" && yarn prettier --write \"src/packages/contracts/generated/infos/**/*.{ts,json}\"",
-    "generate-pancake-swap-tokens": "node ./src/packages/tokens/scripts/generatePancakeSwapTokenRecords.mjs && yarn prettier --write src/packages/tokens/tokenInfos/pancakeSwapTokens/bscMainnet.ts",
+    "generate-pancake-swap-tokens": "./src/packages/tokens/scripts/generatePancakeSwapTokenRecords.ts && yarn prettier --write \"src/packages/tokens/tokenInfos/pancakeSwapTokens/bscMainnet.ts\"",
     "generate-version-file": "genversion --es6 --semi src/constants/version.ts",
     "husky:install": "husky install",
     "postinstall": "npm-run-all --parallel husky:install generate-version-file generate-contracts generate-subgraph-types && yarn generate-pancake-swap-tokens",
